sem5/aiesl/assignment6/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml[html-clean]>=6.0.1",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.1",
    "scikit-learn>=1.7.2",
]
